Moved Json model to another package
