+++ dialect/src/test/java/ydb/jimmer/dialect/model/type/ydbJson/Json.java
@@ -1,4 +1,4 @@
-package ydb.jimmer.dialect.model.type.json;
+package ydb.jimmer.dialect.model.type.ydbJson;
 
 import org.babyfish.jimmer.sql.Serialized;
